removes old finish aggregation method
