+++ tesseract-server/src/handlers/logic_layer/mod.rs
@@ -3,4 +3,4 @@ pub mod shared;
 
 pub use self::aggregate::logic_layer_handler;
 pub use self::aggregate::logic_layer_default_handler;
-pub use self::shared::{Time, TimePrecision, TimeValue, LogicLayerQueryOpt, finish_aggregation};
+pub use self::shared::{Time, TimePrecision, TimeValue, LogicLayerQueryOpt};
